+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,6 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".IndexTest">
-
-        </activity>
         <activity android:name=".IndexPage">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
