+++ app/src/main/java/tk/pluses/plusesprogress/fragments/FragmentDiary.java
@@ -1,26 +1,17 @@
 package tk.pluses.plusesprogress.fragments;
 
+import android.app.ListFragment;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
 
-import tk.pluses.plusesprogress.R;
-
-/**
- * Created by Андрей on 13.07.2017.
- */
-
-public class FragmentDiary extends Fragment implements View.OnClickListener {
+public class FragmentDiary extends ListFragment{
+    String data[] = new String[] { "qwe", "qqq", "sss", "aaa" };
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_diary, container, false);
-    }
-
-    @Override
-    public void onClick (View v) {
-
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        ArrayAdapter<String> adapter = new ArrayAdapter<>(getActivity(),
+                android.R.layout.simple_list_item_1, data);
+        setListAdapter(adapter);
     }
 }
