--- app/src/main/java/tk/pluses/plusesprogress/fragments/ListFragmentTest.java
@@ -1,17 +1,0 @@
-package tk.pluses.plusesprogress.fragments;
-
-import android.app.ListFragment;
-import android.os.Bundle;
-import android.widget.ArrayAdapter;
-
-public class ListFragmentTest extends ListFragment{
-    String data[] = new String[] { "qwe", "qqq", "sss", "aaa" };
-
-    @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(getActivity(),
-                android.R.layout.simple_list_item_1, data);
-        setListAdapter(adapter);
-    }
-}
