--- app/src/main/res/layout/test_page.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/LinearLayout1"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".IndexTest">
-    <fragment
-        android:name="tk.pluses.plusesprogress.fragments.ListFragmentTest"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-    </fragment>
-
-    <android.support.v4.widget.DrawerLayout
-        android:id="@+id/menu_drawer_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <FrameLayout
-            android:id="@+id/menu_frame_layout"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
-
-        <android.support.design.widget.NavigationView
-            android:id="@+id/menu_navigation_view"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_gravity="start"
-            app:headerLayout="@layout/navigate_header"
-            app:menu="@menu/navigate_menu" />
-
-    </android.support.v4.widget.DrawerLayout>
-</LinearLayout>
